update detail product page to use server side rendering
